Added simple backend tester for the app
The test allows the user to input a name and address and submit it to the
database. The database then returns all rows from the test table and the
application displays the data.

This is a good source of proof of concept that will be later removed from
the project. Others can look to see how this is implemented if they are
unsure of how to go about communicating with the server and/or databases.