implemented adding user to database
After a user types in his/her name and presses the 'submit' button, s/he
is added to the 'Users' database. In order for certain foreign keys to
work correctly, there needs to be a defualt hub that each user is
automatically apart of. it has no significance and the user will have no
knowledge of being apart of the hub upon creating a username.