Implemented adding a new hub to the database
When the create hub button is pressed it takes the user to a screen for
him/her to imput a name and 4 digit pin. Upon pressing the Create button,
the user is taken into the queue activity and the information is updated
on the database.

Not all information is updated correctly yet because we do not have
implemented a way of finding out the user's phone id. Until we have that,
the database is receiving random data for the user id.