implemented user phone id for databases
The user's device number is pulled and made the user's primary key within
the users database. Additionally, if a user chooses to create a new hub
then his/her phone id becomes the hub_creator_id. If a user already exists
in the table and s/he tries to create a new username then the old username
is updated without problem.