implemented searching for youtube videos
It is working exactly the way that it was working within the BrianDev
branch. The next task with this will be to send the intent to the
QueueActivity instead of the PlayerActivity which is not a needed activity
and was only there to test if the method would work. We likely want the
QueueActivity to automatically play the song if the song is being added as
the first song within the queue and make sure some sort of adaptive list
is in place for storing the songs within the queue.