implemented backend for search hubs
I think I made a couple changes to the front end as well but nothing
significant. I tested the backend locally, but it has not been tested with
the front end yet. I am pretty sure the front end is not yet ready to
properly send and receive data to/from the server.