synced up front end and back end of hubSearch
The results of the search are correctly displaying in a scrollable screen
on the front end. None of the elements are currently clickable and so we
need to set it up so that they are clickable and upon being clicked, the
user is correctly redirected to the corresponding 'joinHub' page.