Automated UI testing is now possible
Some, possibly all, dependencies have been added for automated UI testing.
We will be using "espresso" which makes automated testing trivial and
allows developers to jump to any activity and test its functionality
instead of having to manually navigate to the activity you want to test.
An example test-case has been written up within the ExampleInstrumentedTest
class under the androidTest directory. The test-case types a string into
the hub search, presses the search button and finally selects an item
within the recycler view. We should create as many test cases for each
activity as we can and keep them short and simple so that they are
thoroughly testing individual parts of the app.