Added hub singleton capabilities
Singleton is workin correctly so that multiple songs can be added to the
queue and remain there upon leaving the queueActivity and entering back to
it. Further implications have come about:

1) Because I didn't want to complete other peoples' tasks, I did not
remove any code and therefore each added song is displayed twice in the
queue.

2) Each new song that is added to the queue is immediately put as the
current song being played (probably just need to set a condition that only
allows the new song to be the currently playing song if the queue is
empty).

3) When the hub coordinator searches for a song s/he leaves the activity
causing the song to stop playing. We will need to make it so the hub
coordinator is able to search for a song without ever leaving the
queueActivity. With the limited space that we have because of the video,
we might need to make it so there is a sliding screen for searching
instead of going to another page for searching.