Added more automated UI testcases
Upon using the automated UI testcases I was able to find a few more places
that were either performing incorrectly or could have been performing
better. I made some cosmetic changes for easier use of the app.

One test case seems to fail because a video tries to load when a brand new
hub is created (we should fix this if it is possible to do so). Another
test case is failing because we do not yet have a check implemented
checking that the user trying to join a hub types the correct pin number
(I went ahead and added a check for making sure that exactly 4 digits were
inserted into the hubpin edit text box).