Creator now searches songs within hub activity
I had to move some things around and it is now a bit messy, but I am
pretty sure that it is working.

We still need to implement removing the song that just finished playing
from the backend to be certain that everything is working correctly.

I had some issues with the threads allowing the youtube player to be
initialized before any songs were retrieved from the backend. I am pretty
sure I have fully debuged the problem and solved the issue, but please let
me know if such a problem arises.