fixed the issue with songId being 0
I have not tested it to see if everything is working perfectly, but
removing the song works correctly and the songs are now reordering
themselves every 5 seconds if necessary. I am not sure if the button color
is correctly updated after the song is moved from one row to another. It
happens so fast it is hard to tell exactly.

I also have a couple of UI changes I forgot to leave on another branch