possible solution for buttons
Made it so that each button can only be pressed once. This way the person
can at least null their vote if they accidently press one of the buttons.

In order for the buttons to behave exactly how we want them to we will
need to be able to increase the pressed button's count while decreasing the other
button's count if it has been pressed before. We currently don't have any php
scripts for decreasing a button so that will not work as of right now.