More UI stuff... again
I updated the view of ViewQueueActivity. It seems to be a little weird and
tries to adjust itself when scrolling to fit the size of the text. I think
we can fix this if we leave a fixed line length or even better have a
scrolling line of text so that it stays on one line.