override the press of the back button
I tried sending the user back to getting started if the back button was
pressed but the app ended up crashing if i tried to do much after. I think
it was because it had memory of the last hub that the user was a part of
so if someone wants to play around with reseting data so the user can
start over feel free.

Currently I have it so that the back button is ignored and nothing happens
if the user presses it while in QueueActivity or ViewQueueActivity. This
is because if the user presses the back button to exit the app and then
renters the app I have found the app crashes randomly. Now the user must
exit the app by pressing the home button (which allows you to reneter the
app where you left off without corrupting data) or by closing the app
completely which makes the user restart from the beggining with no saved
data.