stress test at poll every second
I tried pressing a lot of buttons to see if the app would react poorly to
polling every second. It appears to be just fine, but lags because it
cannot process things as fast as they are pressed. We should go ahead and
demo with this speed because it works while we have such a small number of
users and I am pretty sure it would work with high traffic on the server
as well because of the way aws and the service in android optimally handle
such high traffic.