Allows recording of hub location
A user is prompted with a dialog box to see if s/he wants to allow users
to search for the new hub based on gps location.

if the user selects yes then s/he is taken to the screen to turn on gps.
Upon returning s/he is prompted with a message saying that s/he can
disable gps becuase the location has been saved.

Nothing is implemented on the backend yet and regular users cannot yet
search based on "hubs near me"