implemented user searching hub nearest me
I added a php script that seems to be working correctly on the backend,
but still needs to be tested to see if it plays nice with the front end
which cannot be done until this is merged.

I implemented a user being able to search for a hub nearest him/her on the
front end. It is a trivial approach that just added two buttons to the
search activity and allows the user to change what is displayed by
clicking two different buttons, one displaying the views for
search-by-name and one displaying the views for search-by-location.

I prompt the user for his/her location and then just send this value to
the backend to get the search results and then tell the user s/he can turn
the gps location back off on their phone.

Clicking the button that takes you to the screen that you are already on
does not do anything.