fixed issues, made it look better
I did some stress testing of the gps location stuff and fixed anything
that seemed buggy

I made it so the list dissapears and a loading text is brought up while
the user waits to be prompted about his/her location being saved.

Everything seems to be working as expected on my end. Let me know if
anyone sees something strange happening with it.