changed queue list color view
kept the same colors, but made it so the format is that of what it was
before. I also moved the remove button to the top right. It is a little
overpowering right now so we might want to make it a less vibrant red.
maybe even just a minus sign or just an x sign.